Initial commit
This is simple boilerplate to test some latest versions of tooling.
Unfotunatelly Webpack 5 didn't worked to me, but i will test it later.
Surrently this project is working in developments mode. Production mode is
not tested yet.
Stage: "It works on my PC" Take it or leave it.
Main motivation was to get .scss working. Still researching this area.